major refactoring for valkyrie
