+++ .gitignore
@@ -17,6 +17,7 @@ notebook/static/built/
 notebook/static/*/js/main.min.js*
 notebook/static/lab/*bundle.js
 notebooks/itslive/data/*.nc
+notebooks/Valkyrie/data
 .python-version
 node_modules
 *.py[co]
