+++ Dockerfile
@@ -17,4 +17,6 @@ npm cache clean --force && \
 rm -rf /opt/conda/lib/python3.7/site-packages/awscli/examples \
 && find /opt/conda/ -follow -type f -name '*.a' -delete \
 && find /opt/conda/ -follow -type f -name '*.pyc' -delete \
-&& fix-permissions /home/$NB_USER && fix-permissions $HOME
+&& fix-permissions /home/$NB_USER && fix-permissions $HOME \
+&& jupyter labextension install @jupyter-widgets/jupyterlab-manager jupyter-offlinenotebook \
+bqplot jupyter-matplotlib jupyter-vuetify
