+++ binder/environment.yml
@@ -22,9 +22,12 @@ dependencies:
 - rasterio=1.1.1
 - netcdf4=1.5.3
 - vaex=3.0.0
+- flake8
 - jupyterlab
 - jupyter-offlinenotebook
+- sidecar
 - pip
 - pip:
   - area
+  - git+git://github.com/icesat2py/icepyx@development
   - python-cmr
