+++ notebooks/Valkyrie/.flake8
@@ -1,0 +1,11 @@
+[flake8]
+ignore =
+    F812,   # list comprehension redefines
+    H101    # Use TODO(NAME)
+exclude =
+    .git,
+    __pycache__,
+    docs/source/conf.py,
+    dist
+max-line-length = 120
+max-complexity = 10
